Custom messages handler created
Upon the reviewer advice, I added the KademliaMessage class, which allows creating custom messages with specific structure and fields, it's been implemented inside the SMSKademliaListener to easily create and read messages from and to the network; all the building process has been moved inside the new class. Also, I modified the ResourceExchangeHandler class, now the methods to create and close Requests are generalized. In the next commit I'll add specifications and descriptions for the methods