KademliaBucket ERROR solved
The insertion of a new node inside of the replacementCache WHEN this is full must be handled only if the size of the replacementCache correspond to his maximum size: it can only contain 2 pending nodes, if a new node is added then the oldest one is overwritten. The steps that the process needed were:
1. verify if the actual size is the MAXIMUM one (so that the replacementCache is full)
2. if it is, remove one contact, insert the new one
Tha actual steps were:
1. verify if the actual size is BIGGER then the MAXIMUM one
2. if it is, remove one contact, insert the new one
It means that if the MAXIMUM size of the ReplacementCache was 2, the ReplacementCache could actually accept 3 contact before starting to replace them.