The "Acknowledge" situation
#KadAddResource: added a timer to verifiy the correct accomplishment of the Command
#KadDeleteResource: added a timer to verifiy the correct accomplishment of the Command
Both the above classes needed a way to verify if the node they were trying to contact was able to complete the task they asked him to, also any node who wuold have received a Command from the two classes used to send an acknowledge never asked.
The acknowledge is now used to check if the Commands are completed.
#IdFinderHandler: used to call for the isAlive method even when he was sending the result of a research to the searcher node, with the possibility to create an endless loop